Use getClientRects() rather than getBoundingClientRect()
For an inline elements that wraps to the next line,
getBoundingClientRect() will return a box covering the whole two lines.
The centre of this box may not actually be over the element at all.

getClientRects() returns an array of boxes making up the lines of the
element. So we just click the centre of the first box. For block
elements, it will always return one box, covering the whole element.