Fix test triggering PhantomJS crash.
Closes #163.

Something in driver_spec.rb was causing PhantomJS to crash. This was
reproducible on Ubuntu but not on Fedora. I tried quite hard to track
down the cause but was unable to get a stack trace that came even
remotely close to being useful. Compiling PhantomJS in debug mode made
it go away. Therefore I suspect it was some sort of issue with the JIT
compiler in WebKit, and can only hope this will be solved by WebKit
upgrades in the future.

Fortunately, resetting the driver after each test in driver_spec.rb
makes it go away. This is something we should have been doing anyway,
and adding that actually brought to light a false positive test for a
feature that was in fact broken. So I have fixed that too.