Use any available port
This reverts the default port setting to how it was before the 1.0
release. However the `:port` configuration option is still available for
those who need to set it to something static.

The reason for changing this back is that I have seend EADDRINUSE errors
crop up and in general had second thoughts on whether it was a good idea
to use a static port number as the default. I think it's much better and
less error prone to use any available port by default but to allow the
port number to be configured in the small number of cases where that's
necessary.

Closes #174.

Conflicts:
	README.md