Fix within_window finding window after close/open
Poltergeist was caching the page returned from getPage.  If a popup
window was opened and given a name, then closed, and a new one was
opened with the same name, we couldn't see it because we still had the
old page cached.

This change removes the cache and instead always asks phantomJS for the
page.

Fixes #312