Render full window size if document has no height
When rendering a screenshot with `:full => true` on a document which has
a height of zero, the call fails silently with no screenshot being
created.  This situation can occur with pages that use absolute or fixed
positioning for all content, for example.

Changed to use the `documentElement`'s `clientHeight` & `clientWidth` as
a fallback, which will result in screenshots the size of the viewport in
that case.