Constrain capybara to <v2.3 on our 1.5 support branch
The specs in capybara 2.3 require the 'launchy' gem which as of the 1.5
line of poltergeist is not a dev dependency. I guess it will be added
soon in master but since 2.3 capybara came out, poltergeist v1.5 no
longer passes it's tests.

I also noticed when running the tests that the compiled '.js' files
are different to what has been checked in. Does that mean someone
forgot to check them in along with the 1.5 release or am I doing
something wrong. I decided to leave the generated files out of this
PR for now but can add them in if you like.