Don't allow PhantomJS 2 to be used
This reverts commit 3aec468f9f0f93570b6c58bfd7f67e4cf1a9d360, reversing
changes made to 3786393fe2cc80e3c6ae620b6b4692fa5a76ac87.

In my project a new developer ended up installing PhantomJS 2 when the
rest of us were using 1.9. A bunch of tests failed and it was very
confusing until we realised he was on version 2.

I don't think we can assume that Poltergeist is forward-compatible with
PhantomJS 2:

* The release notes <http://phantomjs.org/release-2.0.html> acknowledge
  that there are various known regressions from 1.x. In general
  PhantomJS 2 is a big rewrite so I expect regressions.

* Poltergeist itself is not currently being tested against PhantomJS 2

I think that the way to add PhantomJS 2 support to Poltergeist should be
in a branch, ensuring that all the tests pass first, rather than just
relaxing the version dependency. This will be much less confusing for
users.